set up everything to work, vercel just isnt running it
