+++ src/pages/Events/Events.jsx
@@ -13,7 +13,7 @@ const Events = () => {
     useEffect(() => {
         const fetchEvents = async () => {
             try {
-                const response = await axios.get('http://localhost:8000/calendar/events');
+                const response = await axios.get('/api/calendar');
                 setEvents(response.data);
             } catch (err) {
                 console.error('Error fetching events:', err);
